Split the FS plugin source into multiple files
The fs.c file has grown very big over the time.
