+++ .gitignore
@@ -6,8 +6,7 @@ tests/error_occurred
 **/*.pyc
 *libblockdev*.tar.gz
 src/lib/plugin_apis/*.[ch]
-src/lib/plugin_apis/btrfs.api
-src/plugins/btrfs.h
+src/plugins/.dirstamp
 BlockDev-*.gir
 BlockDev-*.typelib
 Makefile
@@ -41,6 +40,7 @@ src/utils/blockdev-utils.pc
 **/*.stamp
 **/*.bak
 include/blockdev/*.h
+include/blockdev/fs
 docs/libblockdev-decl-list.txt
 docs/libblockdev-decl.txt
 docs/libblockdev-overrides.txt
