+++ include/blockdev/Makefile.am
@@ -2,8 +2,11 @@ all-local:
 	for header in ${srcdir}/../../src/plugins/*.h; do ln -sf $${header} ./; done
 	for header in ${srcdir}/../../src/utils/*.h; do ln -sf $${header} ./; done
 	for header in ${srcdir}/../../src/lib/*.h; do ln -sf $${header} ./; done
+	mkdir -p fs;
+	cd fs; for header in ${srcdir}/../../../src/plugins/fs/*.h; do ln -sf $${header} ./; done
 
 clean-local:
 	rm -f *.h
+	rm -rf fs
 
 MAINTAINERCLEANFILES = Makefile.in
